+++ multisrc/overrides/wpmangastream/komikcast/src/KomikCast.kt
@@ -1,31 +1,35 @@
 package eu.kanade.tachiyomi.extension.id.komikcast
 
-import eu.kanade.tachiyomi.lib.ratelimit.RateLimitInterceptor
 import eu.kanade.tachiyomi.multisrc.wpmangastream.WPMangaStream
 import eu.kanade.tachiyomi.network.GET
+import eu.kanade.tachiyomi.network.interceptor.rateLimit
 import eu.kanade.tachiyomi.source.model.Filter
 import eu.kanade.tachiyomi.source.model.FilterList
 import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.jsonArray
+import kotlinx.serialization.json.jsonObject
 import okhttp3.Headers
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import okhttp3.OkHttpClient
 import okhttp3.Request
+import org.jsoup.Jsoup
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
+import uy.kohesive.injekt.injectLazy
 import java.util.concurrent.TimeUnit
 
-class KomikCast : WPMangaStream("Komik Cast", "https://komikcast.com", "id") {
+class KomikCast : WPMangaStream("Komik Cast", "https://komikcast.me", "id") {
     // Formerly "Komik Cast (WP Manga Stream)"
     override val id = 972717448578983812
-
-    private val rateLimitInterceptor = RateLimitInterceptor(3)
 
     override val client: OkHttpClient = network.cloudflareClient.newBuilder()
         .connectTimeout(10, TimeUnit.SECONDS)
         .readTimeout(30, TimeUnit.SECONDS)
-        .addNetworkInterceptor(rateLimitInterceptor)
+        .rateLimit(3)
         .build()
 
     override fun headersBuilder(): Headers.Builder = Headers.Builder()
@@ -153,7 +157,23 @@ class KomikCast : WPMangaStream("Komik Cast", "https://komikcast.com", "id") {
     }
 
     override fun pageListParse(document: Document): List<Page> {
-        return document.select("div#chapter_body .main-reading-area img.size-full")
+        var doc = document
+        var cssQuery = "div#chapter_body .main-reading-area img.size-full"
+        val imageListRegex = Regex("chapterImages = (.*) \\|\\|")
+        val imageListMatchResult = imageListRegex.find(document.toString())
+
+        if (imageListMatchResult != null) {
+            val imageListJson = imageListMatchResult.destructured.toList()[0]
+            val imageList = json.parseToJsonElement(imageListJson).jsonObject
+
+            var imageServer = "cdn"
+            if (!imageList.containsKey(imageServer)) imageServer = imageList.keys.first()
+            val imageElement = imageList[imageServer]!!.jsonArray.joinToString("")
+            doc = Jsoup.parse(json.decodeFromString(imageElement))
+            cssQuery = "img.size-full"
+        }
+
+        return doc.select(cssQuery)
             .mapIndexed { i, img -> Page(i, "", img.attr("abs:Src")) }
     }
 
@@ -169,4 +189,6 @@ class KomikCast : WPMangaStream("Komik Cast", "https://komikcast.com", "id") {
         Filter.Header("$name Project List page"),
         ProjectFilter()
     )
+
+    private val json: Json by injectLazy()
 }