+++ multisrc/overrides/wpmangastream/sheamanga/src/SheaManga.kt
@@ -1,7 +1,7 @@
 package eu.kanade.tachiyomi.extension.id.sheamanga
 
-import eu.kanade.tachiyomi.lib.ratelimit.RateLimitInterceptor
 import eu.kanade.tachiyomi.multisrc.wpmangastream.WPMangaStream
+import eu.kanade.tachiyomi.network.interceptor.rateLimit
 import okhttp3.Dns
 import okhttp3.OkHttpClient
 import java.text.SimpleDateFormat
@@ -10,16 +10,15 @@ import java.util.concurrent.TimeUnit
 
 class SheaManga : WPMangaStream(
     "Shea Manga",
-    "http://sheamanga.my.id",
+    "https://sheakomik.com",
     "id",
     dateFormat = SimpleDateFormat("MMMM dd, yyyy", Locale.forLanguageTag("id"))
 ) {
-    private val rateLimitInterceptor = RateLimitInterceptor(4)
 
     override val client: OkHttpClient = network.cloudflareClient.newBuilder()
         .connectTimeout(10, TimeUnit.SECONDS)
         .readTimeout(30, TimeUnit.SECONDS)
-        .addNetworkInterceptor(rateLimitInterceptor)
+        .rateLimit(4)
         .dns(Dns.SYSTEM)
         .build()