+++ src/ru/remanga/src/eu/kanade/tachiyomi/extension/ru/remanga/Remanga.kt
@@ -8,7 +8,6 @@ import MangaDetDto
 import MyLibraryDto
 import PageDto
 import PageWrapperDto
-import PageWrapperDtoNoLimit
 import SeriesWrapperDto
 import TagsDto
 import UserDto
@@ -122,15 +121,9 @@ class Remanga : ConfigurableSource, HttpSource() {
 
     override fun popularMangaParse(response: Response): MangasPage = searchMangaParse(response)
 
-    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/api/titles/last-chapters/?page=$page&count=$count", headers)
+    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/api/search/catalog/?ordering=-chapter_date&count=$count&page=$page", headers)
 
-    override fun latestUpdatesParse(response: Response): MangasPage {
-        val page = json.decodeFromString<PageWrapperDtoNoLimit<LibraryDto>>(response.body!!.string())
-        val mangas = page.content.map {
-            it.toSManga()
-        }
-        return MangasPage(mangas, mangas.isNotEmpty())
-    }
+    override fun latestUpdatesParse(response: Response): MangasPage = searchMangaParse(response)
 
     override fun searchMangaParse(response: Response): MangasPage {
         if (response.request.url.toString().contains("/bookmarks/")) {
@@ -146,9 +139,21 @@ class Remanga : ConfigurableSource, HttpSource() {
             if (preferences.getBoolean(isLib_PREF, false)) {
                 content = content.filter { it.bookmark_type.isNullOrEmpty() }
             }
-            val mangas = content.map {
+
+            var mangas = content.map {
                 it.toSManga()
             }
+
+            if (mangas.isEmpty() && page.props.page < page.props.total_pages && preferences.getBoolean(isLib_PREF, false))
+                mangas = listOf(
+                    SManga.create().apply {
+                        val nextPage = "Пустая страница. Всё в «Закладках»"
+                        title = nextPage
+                        url = nextPage
+                        thumbnail_url = "$baseUrl/icon.png"
+                    }
+                )
+
             return MangasPage(mangas, page.props.page < page.props.total_pages)
         }
     }
@@ -233,11 +238,12 @@ class Remanga : ConfigurableSource, HttpSource() {
 
     private fun parseStatus(status: Int): Int {
         return when (status) {
-            0 -> SManga.COMPLETED
-            1 -> SManga.ONGOING
-            2 -> SManga.ONGOING
-            3 -> SManga.ONGOING
-            5 -> SManga.LICENSED
+            0 -> SManga.COMPLETED // Закончен
+            1 -> SManga.ONGOING // Продолжается
+            2 -> SManga.ON_HIATUS // Заморожен
+            3 -> SManga.ON_HIATUS // Нет переводчика
+            4 -> SManga.ONGOING // Анонс
+            5 -> SManga.LICENSED // Лицензировано
             else -> SManga.UNKNOWN
         }
     }
@@ -252,7 +258,7 @@ class Remanga : ConfigurableSource, HttpSource() {
         return when (age_limit) {
             2 -> "18+"
             1 -> "16+"
-            else -> "0+"
+            else -> ""
         }
     }