+++ src/zh/baozimanhua/src/eu/kanade/tachiyomi/extension/zh/baozimanhua/Baozimanhua.kt
@@ -3,6 +3,7 @@ package eu.kanade.tachiyomi.extension.zh.baozimanhua
 import android.app.Application
 import android.content.SharedPreferences
 import androidx.preference.PreferenceScreen
+import eu.kanade.tachiyomi.AppInfo
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.network.asObservableSuccess
 import eu.kanade.tachiyomi.source.ConfigurableSource
@@ -22,6 +23,8 @@ import org.jsoup.nodes.Element
 import rx.Observable
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
+import java.text.SimpleDateFormat
+import java.util.Locale
 
 class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
 
@@ -29,9 +32,8 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
 
     override val name = "包子漫画"
 
-    private val preferences: SharedPreferences by lazy {
+    private val preferences: SharedPreferences =
         Injekt.get<Application>().getSharedPreferences("source_$id", 0x0000)
-    }
 
     override val baseUrl = "https://${preferences.getString(MIRROR_PREF, MIRRORS[0])}"
 
@@ -40,17 +42,25 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
     override val supportsLatest = true
 
     override val client: OkHttpClient = network.cloudflareClient.newBuilder()
-        .addNetworkInterceptor(BannerInterceptor()).build()
+        .addInterceptor(BannerInterceptor).build()
 
-    override fun chapterListSelector(): String = "div.pure-g[id^=chapter] > div"
+    override fun chapterListSelector() = throw UnsupportedOperationException("Not used.")
 
     override fun chapterListParse(response: Response): List<SChapter> {
         val document = response.asJsoup()
-        val chapters = document.select(chapterListSelector()).map { element ->
-            chapterFromElement(element)
+        return if (document.select(".l-box > .pure-g").size == 1) { // only latest chapters
+            document.select(".l-box > .pure-g > div")
+        } else {
+            // chapters are listed oldest to newest in the source
+            document.select(".l-box > .pure-g[id^=chapter] > div").reversed()
+        }.map { chapterFromElement(it) }.apply {
+            if (!isNewDateLogic) return@apply
+            val date = document.select("em").text()
+            if (date.contains('年')) {
+                this[0].date_upload = date.removePrefix("(").removeSuffix(" 更新)")
+                    .let { DATE_FORMAT.parse(it) }?.time ?: 0L
+            } // 否则要么是没有，要么必然是今天（格式如 "今天 xx:xx", "x小时前", "x分钟前"）可以偷懒直接用默认的获取时间
         }
-        // chapters are listed oldest to newest in the source
-        return chapters.reversed()
     }
 
     override fun chapterFromElement(element: Element): SChapter {
@@ -70,15 +80,12 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
         }
     }
 
-    override fun popularMangaNextPageSelector() = throw java.lang.UnsupportedOperationException("Not used.")
+    override fun popularMangaNextPageSelector(): String? = null
 
-    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/classify", headers)
+    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/classify?page=$page", headers)
 
     override fun popularMangaParse(response: Response): MangasPage {
-        val document = response.asJsoup()
-        val mangas = document.select(popularMangaSelector()).map { element ->
-            popularMangaFromElement(element)
-        }
+        val mangas = super.popularMangaParse(response).mangas
         return MangasPage(mangas, mangas.size == 36)
     }
 
@@ -88,17 +95,9 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
         return popularMangaFromElement(element)
     }
 
-    override fun latestUpdatesNextPageSelector() = throw java.lang.UnsupportedOperationException("Not used.")
+    override fun latestUpdatesNextPageSelector(): String? = null
 
     override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/list/new", headers)
-
-    override fun latestUpdatesParse(response: Response): MangasPage {
-        val document = response.asJsoup()
-        val mangas = document.select(popularMangaSelector()).map { element ->
-            popularMangaFromElement(element)
-        }
-        return MangasPage(mangas, mangas.size == 12)
-    }
 
     override fun mangaDetailsParse(document: Document): SManga {
         return SManga.create().apply {
@@ -117,19 +116,18 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
     }
 
     override fun pageListParse(document: Document): List<Page> {
-        val pages = document.select(".comic-contain > .chapter-img > img").mapIndexed { index, element ->
-            Page(index, imageUrl = element.attr("data-src").trim() + COMIC_IMAGE_SUFFIX)
+        return document.select(".comic-contain > amp-img").mapIndexed { index, element ->
+            Page(index, imageUrl = element.attr("src").trim() + BannerInterceptor.COMIC_IMAGE_SUFFIX)
         }
-        return pages
     }
 
     override fun imageUrlParse(document: Document) = throw UnsupportedOperationException("Not used.")
 
-    override fun searchMangaSelector() = throw java.lang.UnsupportedOperationException("Not used.")
+    override fun searchMangaSelector() = throw UnsupportedOperationException("Not used.")
 
-    override fun searchMangaFromElement(element: Element) = throw java.lang.UnsupportedOperationException("Not used.")
+    override fun searchMangaFromElement(element: Element) = throw UnsupportedOperationException("Not used.")
 
-    override fun searchMangaNextPageSelector() = throw java.lang.UnsupportedOperationException("Not used.")
+    override fun searchMangaNextPageSelector() = throw UnsupportedOperationException("Not used.")
 
     override fun fetchSearchManga(page: Int, query: String, filters: FilterList): Observable<MangasPage> {
         return if (query.startsWith(ID_SEARCH_PREFIX)) {
@@ -155,27 +153,8 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
         return if (query.isNotEmpty()) {
             GET("$baseUrl/search?q=$query", headers)
         } else {
-            lateinit var tag: String
-            lateinit var region: String
-            lateinit var status: String
-            lateinit var start: String
-            filters.forEach { filter ->
-                when (filter) {
-                    is TagFilter -> {
-                        tag = filter.toUriPart()
-                    }
-                    is RegionFilter -> {
-                        region = filter.toUriPart()
-                    }
-                    is StatusFilter -> {
-                        status = filter.toUriPart()
-                    }
-                    is StartFilter -> {
-                        start = filter.toUriPart()
-                    }
-                }
-            }
-            GET("$baseUrl/classify?type=$tag&region=$region&state=$status&filter=$start&page=$page")
+            val parts = filters.filterIsInstance<UriPartFilter>().joinToString("&") { it.toUriPart() }
+            GET("$baseUrl/classify?page=$page&$parts", headers)
         }
     }
 
@@ -204,13 +183,14 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
         StartFilter()
     )
 
-    private open class UriPartFilter(displayName: String, val vals: Array<Pair<String, String>>) :
-        Filter.Select<String>(displayName, vals.map { it.first }.toTypedArray()) {
-        fun toUriPart() = vals[state].second
+    private open class UriPartFilter(name: String, val query: String, val vals: Array<Pair<String, String>>) :
+        Filter.Select<String>(name, vals.map { it.first }.toTypedArray()) {
+        fun toUriPart() = "$query=${vals[state].second}"
     }
 
     private class TagFilter : UriPartFilter(
         "标签",
+        "type",
         arrayOf(
             Pair("全部", "all"),
             Pair("都市", "dushi"),
@@ -281,6 +261,7 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
 
     private class RegionFilter : UriPartFilter(
         "地区",
+        "region",
         arrayOf(
             Pair("全部", "all"),
             Pair("国漫", "cn"),
@@ -292,6 +273,7 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
 
     private class StatusFilter : UriPartFilter(
         "进度",
+        "state",
         arrayOf(
             Pair("全部", "all"),
             Pair("连载中", "serial"),
@@ -301,6 +283,7 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
 
     private class StartFilter : UriPartFilter(
         "标题开头",
+        "filter",
         arrayOf(
             Pair("全部", "*"),
             Pair("ABCD", "ABCD"),
@@ -337,5 +320,14 @@ class Baozimanhua : ParsedHttpSource(), ConfigurableSource {
         private const val MIRROR_PREF_TITLE = "使用镜像网址"
         private const val MIRROR_PREF_SUMMARY = "使用镜像网址。重启软件生效。"
         private val MIRRORS = arrayOf("cn.baozimh.com", "cn.webmota.com")
+
+        private val DATE_FORMAT = SimpleDateFormat("yyyy年MM月dd日", Locale.ENGLISH)
+        private val isNewDateLogic = run {
+            val commitCount = AppInfo.getVersionName().substringAfter('-', "")
+            if (commitCount.isNotEmpty()) // Preview
+                commitCount.toInt() >= 4442
+            else // Stable
+                AppInfo.getVersionCode() >= 81
+        }
     }
 }